requirements.txt
asgiref==3.4.1
Django==3.2.5
python-http-client==3.3.2
pytz==2021.1
sendgrid==6.7.1
sqlparse==0.4.1
starkbank-ecdsa==1.1.1

djangorestframework~=3.12.4
pip~=21.1.3
wheel~=0.34.2
cryptography~=2.8
keyring~=18.0.1
colorama~=0.4.3
PyYAML~=5.3.1
Jinja2~=2.10.1
setuptools~=45.2.0